+++ Sources/MRP/Applications/MVRP/MVRPApplication.swift
@@ -80,7 +80,7 @@ public final class MVRPApplication<P: Port>: BaseApplication, BaseApplicationDel
     }
   }
 
-  public func makeValue(for attributeType: AttributeType, at index: Int) throws -> any Value {
+  public func makeValue(for attributeType: AttributeType, at index: UInt64) throws -> any Value {
     guard let attributeType = MVRPAttributeType(rawValue: attributeType)
     else { throw MRPError.unknownAttributeType }
     switch attributeType {
